simplify code
This unifies the 'no line' / 'existing line' code in Add().
Instead of modifying each line in place, and adding new ones if needed,
this recreates N new lines, and replaces the existing lines (if any) with
the new ones.
This allows to move the low-level line splitting/host deduplication
logic to a small ChunkSlices helper, and makes the higher level code
easier to follow since it's now free of this line splitting logic.