Rework unicity check
This commit groups the code checking for hostname unicity in a single
method instead of removing duplicates from the hostname list in one
place, and then removing entries which are already in the hosts file in
a second place.

lineContains was quite heavyweight, going over the whole file, and then
checking the entry we found is in the right section...
Since we already have filtered lines, we can directly look for the
hostnames in these lines. This has the added benefit of not needing to
differentiate between the 2 cases "there are already lines for this IP"
and "there are no lines for this IP"